Fixes for the automatic update
The auto update crashed with some silly errors.

+++ bunjalloo/arm9/ZipViewer.h
@@ -51,6 +51,6 @@ class ZipViewer: public ButtonListener, public ExtractListener
     ButtonI * m_unzipAndPatch;
     std::vector<CheckBox*> m_checkboxes;
     int m_fileCount;
-    int m_index;
+    size_t m_index;
 };
 #endif
